+++ shared/versions.ts
@@ -1,5 +1,5 @@
-export const APP_VERSION = "2.3.109";
-export const ANDROID_APP_VERSION = "2.3.39";
-export const ANDROID_APK_RELEASE_VERSION = "2.3.109";
+export const APP_VERSION = "2.3.110";
+export const ANDROID_APP_VERSION = "2.3.40";
+export const ANDROID_APK_RELEASE_VERSION = "2.3.110";
 export const AGENT_VERSION = "2.2.86";